requirements.txt
torch
torchvision
onnx
pyserial
pywin32